Create main.go
A slave, a leper or a beggar could act as mystagogues. Also a child can indoctrinate another child. The act itself is trivial, momentary and requires no description. The materials are cork, wax or gum arabic. (The liturgy speaks of silt; this is also often used.) There are no temples specially dedicated to the celebration of this cult, but a ruin, a basement or hall are considered propitious sites.